refactor(web): decompose GitLab setup
