+++ apps/web/src/routes/settings/integrations/gitlab.tsx
@@ -1,98 +1,32 @@
 import { createFileRoute, useNavigate } from '@tanstack/react-router'
+import { zodResolver } from '@hookform/resolvers/zod'
 import { useState } from 'react'
 import { useForm } from 'react-hook-form'
-import { zodResolver } from '@hookform/resolvers/zod'
-import z from 'zod'
 import { toast } from 'sonner'
-import { HugeiconsIcon } from '@hugeicons/react'
-import { Copy01Icon, Refresh01Icon } from '@hugeicons/core-free-icons'
-import type { UseFormReturn } from 'react-hook-form'
 
-import {
-  getActiveInstanceOrRedirect,
-  requireAuthOrRedirect,
-} from '@/lib/instance-context'
+import { GitLabAuthStep } from './-gitlab-auth-step'
+import { GitLabHostStep } from './-gitlab-host-step'
+import { generateWebhookSecret, gitLabSetupSchema } from './-gitlab-setup'
+import { GitLabVerificationStep } from './-gitlab-verification-step'
+import { GitLabWebhookSecretField } from './-gitlab-webhook-secret-field'
+import type { GitLabHostKind, GitLabSetupForm } from './-gitlab-setup'
+import { Alert, AlertDescription } from '@/components/ui/alert'
+import { Button } from '@/components/ui/button'
+import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card'
+import { Form } from '@/components/ui/form'
+import PageHeader from '@/components/page-header'
+import PageLayout from '@/components/page-layout'
 import {
   useExternalAccessNetworkSettings,
   useInstancePreferences,
 } from '@/hooks/use-artifact-storage'
 import { useGitLabStart } from '@/hooks/use-integrations'
+import {
+  getActiveInstanceOrRedirect,
+  requireAuthOrRedirect,
+} from '@/lib/instance-context'
 import { gitLabPublicEndpoints, normalizeGitLabHostUrl } from '@/lib/gitlab-url'
 import { PageMeta } from '@/lib/seo'
-import SetupHint from '@/components/setup-hint'
-import { Button } from '@/components/ui/button'
-import { Alert, AlertDescription } from '@/components/ui/alert'
-import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card'
-import {
-  Form,
-  FormControl,
-  FormDescription,
-  FormField,
-  FormItem,
-  FormLabel,
-  FormMessage,
-} from '@/components/ui/form'
-import { Input } from '@/components/ui/input'
-import { Label } from '@/components/ui/label'
-import PageHeader from '@/components/page-header'
-import PageLayout from '@/components/page-layout'
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from '@/components/ui/select'
-
-const gitLabSetupSchema = z
-  .object({
-    host_url: z
-      .string()
-      .trim()
-      .min(1, 'Host URL is required')
-      .refine(
-        (value) => normalizeGitLabHostUrl(value) !== null,
-        'Use an HTTP(S) host URL only, without a path, query, or credentials.',
-      ),
-    auth_mode: z.enum(['personal_token', 'oauth_app']),
-    webhook_secret: z.string().trim().min(1, 'Webhook secret is required'),
-    access_token: z.string().optional(),
-    client_id: z.string().optional(),
-    client_secret: z.string().optional(),
-  })
-  .superRefine((value, ctx) => {
-    if (value.auth_mode === 'personal_token' && !value.access_token?.trim()) {
-      ctx.addIssue({
-        code: z.ZodIssueCode.custom,
-        message: 'Access token is required',
-        path: ['access_token'],
-      })
-    }
-
-    if (value.auth_mode === 'oauth_app' && !value.client_id?.trim()) {
-      ctx.addIssue({
-        code: z.ZodIssueCode.custom,
-        message: 'Client ID is required',
-        path: ['client_id'],
-      })
-    }
-
-    if (value.auth_mode === 'oauth_app' && !value.client_secret?.trim()) {
-      ctx.addIssue({
-        code: z.ZodIssueCode.custom,
-        message: 'Client secret is required',
-        path: ['client_secret'],
-      })
-    }
-  })
-
-type GitLabSetupForm = z.infer<typeof gitLabSetupSchema>
-type GitLabHostKind = 'gitlab_com' | 'self_managed'
-
-function generateWebhookSecret(): string {
-  const bytes = crypto.getRandomValues(new Uint8Array(24))
-  return `oore_${Array.from(bytes, (byte) => byte.toString(16).padStart(2, '0')).join('')}`
-}
 
 export const Route = createFileRoute('/settings/integrations/gitlab')({
   staticData: { breadcrumbLabel: 'GitLab' },
@@ -112,7 +46,6 @@ function GitLabSetupPage() {
   const remoteEnabled = preferences?.preferences.runtime_mode === 'remote'
   const [webhookSecret] = useState(generateWebhookSecret)
   const [hostKind, setSelectedHostKind] = useState<GitLabHostKind>('gitlab_com')
-
   const form = useForm<GitLabSetupForm>({
     resolver: zodResolver(gitLabSetupSchema),
     mode: 'onBlur',
@@ -125,7 +58,6 @@ function GitLabSetupPage() {
       client_secret: '',
     },
   })
-
   const authMode = form.watch('auth_mode')
   const hostUrl = form.watch('host_url')
   const normalizedHostUrl =
@@ -139,6 +71,7 @@ function GitLabSetupPage() {
     if (!remoteEnabled) return
     const submittedHostUrl = normalizeGitLabHostUrl(data.host_url)
     if (!submittedHostUrl) return
+
     startMutation.mutate(
       {
         host_url: submittedHostUrl,
@@ -159,32 +92,29 @@ function GitLabSetupPage() {
       },
       {
         onSuccess: (response) => {
+          const displayName = response.integration.display_name ?? 'GitLab'
           if (response.integration.status === 'inactive') {
             toast.message(
-              `Host verified. Authorize ${response.integration.display_name ?? 'GitLab'} to complete setup.`,
+              `Host verified. Authorize ${displayName} to complete setup.`,
             )
           } else {
-            toast.success(
-              `Source verified: ${response.integration.display_name ?? 'GitLab'}`,
-            )
+            toast.success(`Source verified: ${displayName}`)
           }
           void navigate({
             to: '/settings/integrations/$integrationId',
             params: { integrationId: response.integration.id },
           })
         },
-        onError: (err) => {
-          toast.error(`Failed to connect GitLab: ${err.message}`)
-        },
+        onError: (err) =>
+          toast.error(`Failed to connect GitLab: ${err.message}`),
       },
     )
   }
 
   function normalizeHostUrl() {
     const normalized = normalizeGitLabHostUrl(form.getValues('host_url'))
-    if (normalized) {
+    if (normalized)
       form.setValue('host_url', normalized, { shouldValidate: true })
-    }
   }
 
   function selectHostKind(value: GitLabHostKind | null) {
@@ -221,7 +151,6 @@ function GitLabSetupPage() {
         description="Connect GitLab.com or a self-managed GitLab host for repository discovery and webhook-triggered builds."
         back={{ to: '/settings/integrations', label: 'Sources' }}
       />
-
       <Card>
         <CardHeader>
           <CardTitle className="text-sm font-medium uppercase tracking-wider text-muted-foreground">
@@ -242,148 +171,19 @@ function GitLabSetupPage() {
               onSubmit={form.handleSubmit(handleSubmit)}
               className="space-y-6"
             >
-              <section className="space-y-4">
-                <div>
-                  <p className="text-sm font-medium uppercase tracking-wider text-muted-foreground">
-                    1. Choose GitLab host
-                  </p>
-                  <p className="mt-1 text-sm text-muted-foreground">
-                    GitLab.com and self-managed GitLab use the same connection
-                    flow.
-                  </p>
-                </div>
-                <div className="space-y-2">
-                  <Label htmlFor="gitlab-host">GitLab host</Label>
-                  <Select
-                    value={hostKind}
-                    onValueChange={selectHostKind}
-                    items={{
-                      gitlab_com: 'GitLab.com',
-                      self_managed: 'Self-managed GitLab',
-                    }}
-                  >
-                    <SelectTrigger id="gitlab-host">
-                      <SelectValue />
-                    </SelectTrigger>
-                    <SelectContent>
-                      <SelectItem value="gitlab_com">GitLab.com</SelectItem>
-                      <SelectItem value="self_managed">
-                        Self-managed GitLab
-                      </SelectItem>
-                    </SelectContent>
-                  </Select>
-                </div>
-                {hostKind === 'self_managed' ? (
-                  <FormField
-                    control={form.control}
-                    name="host_url"
-                    render={({ field }) => (
-                      <FormItem>
-                        <FormLabel>Self-managed host URL</FormLabel>
-                        <FormControl>
-                          <Input
-                            {...field}
-                            placeholder="https://gitlab.example.com"
-                            onBlur={() => {
-                              field.onBlur()
-                              normalizeHostUrl()
-                            }}
-                          />
-                        </FormControl>
-                        <FormDescription>
-                          Host origin only. Oore normalizes a trailing slash; do
-                          not include <code>/api/v4</code> or a group path.
-                        </FormDescription>
-                        <FormMessage />
-                      </FormItem>
-                    )}
-                  />
-                ) : (
-                  <SetupHint
-                    title="GitLab.com selected"
-                    items={[
-                      'Oore will connect to https://gitlab.com. Choose self-managed only when your GitLab has a different host URL.',
-                    ]}
-                  />
-                )}
-              </section>
-
-              <section className="space-y-4 border-t border-border/60 pt-6">
-                <div>
-                  <p className="text-sm font-medium uppercase tracking-wider text-muted-foreground">
-                    2. Authenticate
-                  </p>
-                  <p className="mt-1 text-sm text-muted-foreground">
-                    Choose the credential model that fits this GitLab source.
-                  </p>
-                </div>
-                <FormField
-                  control={form.control}
-                  name="auth_mode"
-                  render={({ field }) => (
-                    <FormItem>
-                      <FormLabel>Authentication method</FormLabel>
-                      <Select
-                        value={field.value}
-                        onValueChange={field.onChange}
-                        items={{
-                          personal_token: 'Personal Access Token',
-                          oauth_app: 'OAuth Application',
-                        }}
-                      >
-                        <FormControl>
-                          <SelectTrigger>
-                            <SelectValue />
-                          </SelectTrigger>
-                        </FormControl>
-                        <SelectContent>
-                          <SelectItem value="personal_token">
-                            Personal Access Token
-                          </SelectItem>
-                          <SelectItem value="oauth_app">
-                            OAuth Application
-                          </SelectItem>
-                        </SelectContent>
-                      </Select>
-                      <FormDescription>
-                        Personal access tokens are fastest for one account and
-                        are verified before saving. OAuth keeps user
-                        authorization in GitLab and is better for a shared
-                        source; it requires one additional authorization after
-                        saving.
-                      </FormDescription>
-                      <FormMessage />
-                    </FormItem>
-                  )}
-                />
-
-                <GitLabCredentialsFields
-                  form={form}
-                  authMode={authMode}
-                  hostUrl={normalizedHostUrl}
-                  callbackUrl={callbackUrl}
-                />
-              </section>
-
-              <section className="space-y-3 border-t border-border/60 pt-6">
-                <div>
-                  <p className="text-sm font-medium uppercase tracking-wider text-muted-foreground">
-                    3. Verify source connection
-                  </p>
-                  <p className="mt-1 text-sm text-muted-foreground">
-                    Oore checks the selected GitLab host before it stores this
-                    source.
-                  </p>
-                </div>
-                <Alert>
-                  <AlertDescription>
-                    {authMode === 'personal_token'
-                      ? 'Saving verifies this token against your GitLab account and discovers accessible projects. If it fails, create a new token with the listed read-only scopes.'
-                      : 'Saving checks that this GitLab host is reachable. Then authorize the saved OAuth application on GitLab; the source becomes active only after the callback returns.'}
-                  </AlertDescription>
-                </Alert>
-              </section>
-
+              <GitLabHostStep
+                form={form}
+                hostKind={hostKind}
+                onHostKindChange={selectHostKind}
+                onHostUrlBlur={normalizeHostUrl}
+              />
+              <GitLabAuthStep
+                form={form}
+                authMode={authMode}
+                hostUrl={normalizedHostUrl}
+                callbackUrl={callbackUrl}
+              />
+              <GitLabVerificationStep authMode={authMode} />
               <section className="space-y-4 border-t border-border/60 pt-6">
                 <div>
                   <p className="text-sm font-medium uppercase tracking-wider text-muted-foreground">
@@ -401,7 +201,6 @@ function GitLabSetupPage() {
                   onRegenerate={replaceWebhookSecret}
                 />
               </section>
-
               <Button
                 type="submit"
                 disabled={
@@ -425,172 +224,5 @@ function GitLabSetupPage() {
         </CardContent>
       </Card>
     </PageLayout>
-  )
-}
-
-function GitLabWebhookSecretField({
-  form,
-  webhookUrl,
-  onCopy,
-  onRegenerate,
-}: {
-  form: UseFormReturn<GitLabSetupForm>
-  webhookUrl: string
-  onCopy: () => void
-  onRegenerate: () => void
-}) {
-  return (
-    <FormField
-      control={form.control}
-      name="webhook_secret"
-      render={({ field }) => (
-        <FormItem>
-          <FormLabel>Webhook secret</FormLabel>
-          <div className="flex gap-2">
-            <FormControl>
-              <Input type="password" {...field} />
-            </FormControl>
-            <Button
-              type="button"
-              variant="outline"
-              size="icon"
-              onClick={onCopy}
-              aria-label="Copy webhook secret"
-              title="Copy webhook secret"
-            >
-              <HugeiconsIcon icon={Copy01Icon} />
-            </Button>
-            <Button
-              type="button"
-              variant="outline"
-              size="icon"
-              onClick={onRegenerate}
-              aria-label="Generate a new webhook secret"
-              title="Generate a new webhook secret"
-            >
-              <HugeiconsIcon icon={Refresh01Icon} />
-            </Button>
-          </div>
-          <FormDescription>
-            Generated securely in this browser. Copy it into GitLab; Oore
-            encrypts it when this source is saved.
-          </FormDescription>
-          <SetupHint
-            title="Webhook setup in GitLab"
-            items={[
-              <span>
-                URL: <code>{webhookUrl}</code>
-              </span>,
-              'Project Settings -> Webhooks -> Secret token. Enable Push events (and Merge request events if your pipeline uses them).',
-            ]}
-          />
-          <FormMessage />
-        </FormItem>
-      )}
-    />
-  )
-}
-
-function GitLabCredentialsFields({
-  form,
-  authMode,
-  hostUrl,
-  callbackUrl,
-}: {
-  form: UseFormReturn<GitLabSetupForm>
-  authMode: GitLabSetupForm['auth_mode']
-  hostUrl: string
-  callbackUrl: string
-}) {
-  if (authMode === 'personal_token') {
-    return (
-      <FormField
-        control={form.control}
-        name="access_token"
-        render={({ field }) => (
-          <FormItem>
-            <FormLabel>Access token</FormLabel>
-            <FormControl>
-              <Input type="password" {...field} placeholder="glpat-..." />
-            </FormControl>
-            <SetupHint
-              title="Required GitLab PAT scopes"
-              items={[
-                <span>
-                  Select <code>read_user</code>, <code>read_api</code>, and{' '}
-                  <code>read_repository</code>.
-                </span>,
-                <span>
-                  Do not select full <code>api</code> unless you are testing a
-                  future write-capable GitLab feature.
-                </span>,
-                <span>
-                  Create it at{' '}
-                  <code>{hostUrl}/-/user_settings/personal_access_tokens</code>.
-                </span>,
-              ]}
-            />
-            <FormMessage />
-          </FormItem>
-        )}
-      />
-    )
-  }
-
-  return (
-    <>
-      <FormField
-        control={form.control}
-        name="client_id"
-        render={({ field }) => (
-          <FormItem>
-            <FormLabel>Client ID</FormLabel>
-            <FormControl>
-              <Input {...field} placeholder="Application ID" />
-            </FormControl>
-            <FormDescription>
-              Create an OAuth application on {hostUrl} and paste its Application
-              ID here.
-            </FormDescription>
-            <FormMessage />
-          </FormItem>
-        )}
-      />
-      <FormField
-        control={form.control}
-        name="client_secret"
-        render={({ field }) => (
-          <FormItem>
-            <FormLabel>Client secret</FormLabel>
-            <FormControl>
-              <Input
-                type="password"
-                {...field}
-                placeholder="Application secret"
-              />
-            </FormControl>
-            <FormDescription>
-              Use the Secret from the same GitLab OAuth application.
-            </FormDescription>
-            <SetupHint
-              title="OAuth callback"
-              items={[
-                <span>
-                  Register this redirect URI in GitLab:{' '}
-                  <code>{callbackUrl}</code>
-                </span>,
-                <span>
-                  Request only <code>read_api</code> and{' '}
-                  <code>read_repository</code>; Oore does not request write
-                  scopes for this source.
-                </span>,
-                'Save this source, then choose Authorize on GitLab from its source details page.',
-              ]}
-            />
-            <FormMessage />
-          </FormItem>
-        )}
-      />
-    </>
   )
 }
